Add commonly used Qt5 packages
Many applications in Fedora use the Qt5 toolkit. Add Qt5 libraries
that are used by at least 20 applications in Fedora. This list is
similar (but not identical) to the list of Qt5 libraries installed
by the qt5 metapackage and by the org.kde.Platform runtime.
(In particular, qt3d is excluded because it's large and not used
by any graphical application in Fedora.)