Add flatpak-runtime-config to the srpm-buildroot group
When we are building a SRPM that in a module that buildrequires
flatpak-runtime, already built packages will be in the repositories
that are used to install the "srpm-build" packages. We might end
up with libraries in /app/lib[64] because of this, so we need
flatpak-runtime-config.

For example, the libreoffice build rebuilds libssh - this is a
dependency of curl, so having a rebuilt libssh without having
/app/lib[64] in the path breaks 'rhpkg-simple sources'.

Also include flatpak-rpm-macros - we don't expect that any macros
in there would change how a SRPM is built, but if they *do* change
how a SPRM is built, then we want them installed when building the
SRPM.