Add back libcanberra
It was dropped in upstream GNOME 3.38 runtime, but many of our packages
still depend on it (12 flatpaks as of right now).

Add it back here for F33 as I ran into a build failure building libtdb
with /app prefix and it really makes sense to share it as we have so
many users.