Don't fail on empty dependencies from fedmod resolve-deps
The rich dependencies are not at the moment handled by fedmod's
resolve-deps. Which could lead to fedmod failing when asking it for JSON
output and in other cases - reported as [0]. Owen prepared a patch[1] that
just returns an empty dependency in case of rich dependencies, but the
code in generate-runtime-report.py doesn't count with empty value there
so it needs to be adapted.

[0] - https://pagure.io/modularity/fedmod/issue/105
[1] - https://pagure.io/modularity/fedmod/issue/105#comment-795783