Add google-noto-serif-vf-fonts to runtime
See https://github.com/fedora-silverblue/issue-tracker/issues/305
@mcatanzaro and I believe that this package not being in the
runtime, combined with some weirdness to do with fontconfig
caches, is likely the cause of help not working in various app
flatpaks. The weirdness is being poked at, but it seems in any
case a good idea to include Fedora's official default serif font
in the runtime.

Signed-off-by: Adam Williamson <awilliam@redhat.com>