Add google-noto-sans-mono-vf-fonts to runtime
See https://github.com/fedora-silverblue/issue-tracker/issues/299
As with the recent addition of the serif fonts, this is necessary
for flatpaks to show monospace fonts correctly in some cases.
We believe fontconfig should manage to fall back correctly to a
different font in these cases, but currently it does not, so it
makes sense to add the Fedora default font to the runtime as a
mitigation.

Signed-off-by: Adam Williamson <awilliam@redhat.com>