Drop Qt 5
... as discussed in the latest Flatpak SIG meeting.

The plan is to use the same strategy as with Qt 6 (that has never been
part of the runtime): Move both Qt 5 and Qt 6 to flatpak-common, which
should make them seamlessly available to app flatpaks.

By dropping Qt 5, we save 64 MB in the runtime, while at the same time
making it easier for app flatpaks to make sure that they are updated in
lock step with the Qt version that they were built against, which is
important for some that use private symbols from Qt.

The tradeoff is that while we save space in the default install
(Silverblue doesn't ship any apps by default that use Qt 5), we get app
flatpaks that take up more space.

This is fairly experimental and if the flatpak-common plan doesn't work
out, we may end up adding both Qt 5 and Qt 6 to the runtime again.
Adding doesn't break ABI so we can do that during the F38 life time if
needed.