util.py: Use flatpak-module-depchase fetch-metadata --print-location
To avoid hardcoding cache locations, use a new flag for
flatpak-module-depchase to print out the repodata/ location.

Support for multiple repositories is removed - it's possible we'll need
it again in the future for EL runtimes, but then flatpak-module-depchase
will need a different way to specify the repositories than '--tag=TAG'.