Update to FD.o 23.08 / GNOME 45
The major change in these versions is the addition of perl to
org.gnome.Platform (not just Sdk), but that was by accident and should
be fixed upstream.  Also, gtksourceview4 was removed upstream, but this
keeps it in for F39; in F40 this should be removed from the runtime, and
will need to be built in /app for a number of flatpaks.

These versions require a lot more filtering for unwanted components,
including system services such as avahi and systemd.

Many libraries are now in sync with F39 versions, but their renames are
commented out (rather than removed) to make it easier to restore the
overrides when they go out of sync again in the future.

Upstream has switched from libappstream-glib's appstream-compose command
to appstreamcli-compose (from the appstream-compose subpackage).  Since
Fedora's tooling still uses the former, it is forced to remain for now,
but the tooling will likely need to be updated in a future version.
Whenever that happens, rpm (or at a minimum, lua-libs) will need to
remain to avoid the koji buildroot dependency issue.