Add back Qt 5 and Qt 6
This adds back Qt 5 that we temporarily dropped from the F38 runtime as
an experiment, and adds Qt 6. Compared to the previous Qt package set
that we had in F37, we no longer try to exactly match Workstation Qt
packages but instead just add whatever is needed by actual app flatpaks.

The F38 experiment worked well, but now with the new flatpak build
system in F39 the /app-built Qt modules started causing build issues for
KDE flatpaks (as f39-flatpak-app build root is shared). This moves them
to the runtime to avoid causing issues for KDE builds.