Add sgml-common, libgccjit
opensp, a dependency of libofx used by financial applications, requires
sgml-common. sgml-common and xml-common are from the same SRPM, and
having the latter in /app is likely to break the buildroot.  opensp will
require modification to use a fixed /etc and /usr prefix, but that seems
simpler.

emacs requires libgccjit, which is part of and requires gcc. This is
another reason to use flatpak-sdk as its base (the flathub flatpak is
also based on the SDK), which is now possible with the new tooling.