Allow for passing more generic arguments to flatpak-module-depchase
Instead of hardcoding a TAG_ARG that we use in various places, instead
have a central function to call flatpak-module-depchase, and allow
for arbitrary repository arguments.

Since the arbitrary arguments could allow for multiple repositories,
make the "get_repo_cacheable()" infrastructure allow that, as long
as the "cacheable" is a map - which it always is for us currently.