Dynamically determine which packages are CRB and should be excluded
Instead of maintaining a list of CRB packages that need to be
excluded from the Platform, use a different set of repositories when
resolving files and dependencies for the platform or SDK. This maintains
many of the same exclusions, but some packages that have made
their way out of the SDK are added to the Platform. Also, we no
longer try to exclude -devel packages from the SDK if the corresponding
libraries aren't in the Platform - this turned out to make virtually
no difference.