Enable dynamic ld.so.cache generation
Using LD_LIBRARY_PATH for adding to the default library search path is not
always reliable, as sometimes subprocesses may be launched without this
environment variable, which may cause it to fail if it needs libraries
outside of the runtime.

Instead, we should follow the lead of Flathub flatpaks, and use dynamic
ld.so.cache generation instead.  This automatically creates (or updates) a
cache file in ~/.var/app/$APP_ID/.ld.so/ based on the runtime, app, and any
extensions, and works even for subprocesses with a limited environment.

To implement this, in addition to the changes here, flatpak-runtime-config
includes a more extensive ld.so.conf.d snippet:

https://src.fedoraproject.org/rpms/flatpak-runtime-config/c/641ea400561c1c79405c4e88aa248c8add914e19?branch=f41

This is necessary because the feature requires ldconfig in /bin (where we
still have a separate /sbin), and the code is not designed to handle
/etc/ld.so.conf.d/*.conf.  This also allows %_sysconfdir/ld.so.conf.d to
work for both RPM builds and flatpak apps rather than having to defining
--env=LD_LIBRARY_PATH in the app too.

https://github.com/flatpak/flatpak/blob/1.15.12/common/flatpak-run.c#L2663
https://github.com/flatpak/flatpak/blob/1.15.12/common/flatpak-run.c#L3083

Fixes: rhbz#2272376